requirements.txt
Pillow==9.0.1
pytesseract==0.3.10
